requirements.txt
streamlit==0.89.0
tensorflow==2.6.0
numpy==1.19.5
Pillow==8.3.2
